- Project structure refactored

+++ src/main/java/com/bestgroup/core/domain/Entity.java
@@ -4,15 +4,15 @@ import java.util.Date;
 
 public class Entity {
 	
-	protected int id;
+	protected Long id;
 	protected Date createdAt;
 	protected Date updatedAt;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 	
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
